app/src/androidTest/java/com/example/cmput301project/MainActivityTest.java
package com.example.cmput301project;


import static org.hamcrest.CoreMatchers.not;
import static org.junit.Assert.assertEquals;
import static org.junit.Assert.fail;
import static org.mockito.Mockito.mock;
import static androidx.test.espresso.Espresso.onView;
import static androidx.test.espresso.action.ViewActions.click;
import static androidx.test.espresso.assertion.ViewAssertions.matches;
import static androidx.test.espresso.intent.Intents.intended;
import static androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent;
import static androidx.test.espresso.matcher.RootMatchers.withDecorView;
import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
import static androidx.test.espresso.matcher.ViewMatchers.isRoot;
import static androidx.test.espresso.matcher.ViewMatchers.withId;
import static androidx.test.espresso.matcher.ViewMatchers.withText;

import static java.lang.Thread.sleep;

import android.view.View;

import androidx.test.core.app.ActivityScenario;
import androidx.test.espresso.Espresso;
import androidx.test.espresso.action.ViewActions;
import androidx.test.espresso.intent.Intents;
import androidx.test.ext.junit.rules.ActivityScenarioRule;
import androidx.test.ext.junit.runners.AndroidJUnit4;

import com.example.cmput301project.activities.LoginActivity;
import com.example.cmput301project.activities.MainActivity;
import com.example.cmput301project.activities.SignUpActivity;
import com.example.cmput301project.itemClasses.Item;
import com.google.firebase.auth.FirebaseAuth;
import com.google.firebase.auth.FirebaseUser;

import org.junit.Before;
import org.junit.Rule;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mock;

import java.lang.reflect.Method;
import java.util.ArrayList;


@RunWith(AndroidJUnit4.class)
public class MainActivityTest {

    // Rule to launch MainActivity for testing
    @Rule
    public ActivityScenarioRule<MainActivity> activityScenarioRule = new ActivityScenarioRule<>(MainActivity.class);

    // Test for initialization
    @Test
    public void testInitialization() {

    }

    // Test AddItem functionality
    @Test
    public void testAddItem() {

    }

    // Test Item Deletion
    @Test
    public void testItemDeletion() {

    }

    @Test
    public void onCreate() {
    }

    @Test
    public void onOKPressed() {
    }

    @Test
    public void updateTotalCost() {
    }

    @Test
    public void editItem() {
    }

    @Test
    public void viewItem() {
    }

    @Test
    public void onDeletePressed() {
    }

    @Test
    public void onFiltersSaved() {
    }

    @Test
    public void onFiltersCleared() {
    }

    @Test
    public void onItemEdited() {
    }

    @Test
    public void updateTotalCostAfterEdit() {
    }

    // Add more test methods for other functionalities following a similar pattern

}
